+++ Cargo.toml
@@ -35,7 +35,6 @@ members = [
     "module-system/module-implementations/sov-blob-storage",
     "module-system/module-implementations/sov-evm",
     "module-system/module-implementations/sov-ibc",
-    "module-system/module-implementations/sov-ibc-transfer",
     "module-system/module-implementations/sov-prover-incentives",
     "module-system/module-implementations/sov-attester-incentives",
     "module-system/module-implementations/sov-sequencer-registry",
