--- module-system/module-implementations/sov-ibc-transfer/Cargo.toml
@@ -1,34 +1,0 @@
-[package]
-name = "sov-ibc-transfer"
-authors = { workspace = true }
-edition = { workspace = true }
-homepage = { workspace = true }
-license = { workspace = true }
-repository = { workspace = true }
-rust-version = { workspace = true }
-version = { workspace = true }
-readme = "README.md"
-publish = false
-resolver = "2"
-
-[dependencies]
-anyhow = { workspace = true }
-ibc = { workspace = true }
-sov-bank = { path = "../sov-bank", default-features = false }
-sov-modules-api = { path = "../../sov-modules-api", default-features = false, features = ["macros"] }
-sov-state = { path = "../../sov-state", default-features = false }
-sov-rollup-interface = { path = "../../../rollup-interface" }
-schemars = { workspace = true, optional = true }
-serde = { workspace = true, optional = true }
-serde_json = { workspace = true, optional = true }
-thiserror = { workspace = true }
-uint = "0.9"
-borsh = { workspace = true, features = ["rc"] }
-
-[dev-dependencies]
-tempfile = { workspace = true }
-
-[features]
-default = ["native"]
-serde = ["dep:serde", "dep:serde_json"]
-native = ["serde", "sov-modules-api/native", "schemars"]
