--- module-system/module-implementations/sov-ibc-transfer/README.md
@@ -1,1 +1,0 @@
-Implements the IBC Router (ICS 26).
