+++ module-system/module-implementations/sov-ibc/src/applications/mod.rs
@@ -1,0 +1,1 @@
+pub mod transfer;
