+++ module-system/module-implementations/sov-ibc/src/applications/transfer/mod.rs
@@ -17,11 +17,11 @@ pub struct TransferConfig {}
 pub struct Transfer<C: Context> {
     /// Address of the module.
     #[address]
-    pub address: C::Address,
+    pub(crate) address: C::Address,
 
     /// Reference to the Bank module.
     #[module]
-    pub bank: sov_bank::Bank<C>,
+    pub(crate) bank: sov_bank::Bank<C>,
 
     /// Keeps track of the address of each token we minted by token denom.
     #[state]
@@ -38,7 +38,7 @@ pub struct Transfer<C: Context> {
     /// 1. when tokens are escrowed, save the mapping `denom -> token address`
     /// 2. when tokens are unescrowed, lookup the token address by `denom`
     #[state]
-    pub escrowed_tokens: StateMap<String, C::Address>,
+    pub(crate) escrowed_tokens: StateMap<String, C::Address>,
 }
 
 impl<C: Context> Module for Transfer<C> {
